This resolves a build failure caused by Qt 6 marking QFile::open() as [[nodiscard]] #32
With -Werror=unused-result, calling open() without checking it's return value causes the build to fail.

To fix this, the code in GameConfiguration::save() has been updated to use the return value of QFile::open() rather than ignoring it.
The logic now checks whether the source file opened successfully before attempting to read from it.

Going forward, an else condition could be used to handle what happens when a file does not open properly.